requirements-package.txt
# These are the library requirements
mlflow
algorithmia
gitpython
jinja2
